refactor(core): move provider-owned local runtime services
Move concrete Weixin, browser launch, local shell resolution, and Computer Use system primitives into service owner crates while preserving core compatibility facades.

Update boundary rules and local module guidance for the new provider ownership.
