+++ src/crates/assembly/core/src/agentic/tools/implementations/computer_use_actions.rs
@@ -10,9 +10,11 @@ use crate::agentic::tools::computer_use_host::{
     InteractiveViewOpts, VisualClickParams, VisualMarkViewOpts,
 };
 use crate::agentic::tools::framework::{Tool, ToolResult, ToolUseContext};
-use crate::util::elapsed_ms_u64;
 use crate::util::errors::{BitFunError, BitFunResult};
-use crate::util::process_manager;
+use bitfun_services_core::system::{
+    truncate_with_marker, LocalSystemActionError, LocalSystemActionErrorKind, LocalSystemProvider,
+    RunScriptRequest,
+};
 use serde_json::{json, Value};
 
 use super::control_hub::{err_response, ControlHubError, ErrorCode};
@@ -260,18 +262,8 @@ impl ComputerUseActions {
                     _ => vec!["return".to_string()],
                 };
 
-                if let Err(e) = clipboard_write(text).await {
-                    return Ok(err_response(
-                        "desktop",
-                        "paste",
-                        ControlHubError::new(
-                            ErrorCode::NotAvailable,
-                            format!("Clipboard write failed: {}", e),
-                        )
-                        .with_hint(
-                            "Fall back to type_text or check that wl-clipboard / xclip is installed (Linux only)",
-                        ),
-                    ));
+                if let Err(e) = LocalSystemProvider::new().clipboard_write_text(text).await {
+                    return Ok(local_system_error_response("desktop", "paste", e));
                 }
 
                 let paste_chord = match std::env::consts::OS {
@@ -1297,48 +1289,6 @@ impl ComputerUseActions {
 
     // ── Browser domain ─────────────────────────────────────────────────
 
-    ///   try in order: `gtk-launch <name>` (uses `.desktop` files), then a
-    ///   direct exec of the lower-cased name (handles `firefox`, `code`, etc.),
-    ///   and finally fall back to `xdg-open` so callers passing a URL/path by
-    ///   accident still work. The dispatcher in `handle_system` is aware of
-    ///   this fallback chain.
-    fn platform_open_command(app_name: &str) -> (String, Vec<String>) {
-        #[cfg(target_os = "macos")]
-        {
-            (
-                "open".to_string(),
-                vec!["-a".to_string(), app_name.to_string()],
-            )
-        }
-        #[cfg(target_os = "windows")]
-        {
-            (
-                "cmd".to_string(),
-                vec![
-                    "/C".to_string(),
-                    "start".to_string(),
-                    "".to_string(),
-                    app_name.to_string(),
-                ],
-            )
-        }
-        #[cfg(target_os = "linux")]
-        {
-            // Probe in order of correctness; the first executable on PATH wins.
-            // `gtk-launch` is the canonical way to start a desktop application
-            // by its .desktop id; if not present we fall back to a direct exec.
-            if which_exists("gtk-launch") {
-                ("gtk-launch".to_string(), vec![app_name.to_string()])
-            } else {
-                (app_name.to_string(), vec![])
-            }
-        }
-        #[cfg(not(any(target_os = "macos", target_os = "windows", target_os = "linux")))]
-        {
-            ("open".to_string(), vec![app_name.to_string()])
-        }
-    }
-
     // ── System domain ──────────────────────────────────────────────────
 
     pub(crate) async fn handle_system(
@@ -1402,91 +1352,24 @@ impl ComputerUseActions {
                     }
                 }
 
-                // Build the platform-specific launch attempt list. On Linux
-                // we try multiple strategies in order so the model doesn't
-                // need to know whether the user has gtk-launch installed.
-                let attempts: Vec<(String, Vec<String>)> = {
-                    let primary = Self::platform_open_command(app_name);
-                    #[cfg(target_os = "linux")]
-                    {
-                        let mut v = vec![primary];
-                        // Fallback 1: direct exec of the lowercase name (handles
-                        // `firefox`, `code`, `gnome-terminal`, etc. when the
-                        // exec name matches the app name).
-                        let lower = app_name.to_lowercase();
-                        if v.iter().all(|(c, _)| c != &lower) {
-                            v.push((lower, vec![]));
-                        }
-                        // Fallback 2: xdg-open — last-ditch, mostly for paths/URLs
-                        // erroneously passed as app_name.
-                        v.push(("xdg-open".to_string(), vec![app_name.to_string()]));
-                        v
-                    }
-                    #[cfg(not(target_os = "linux"))]
-                    {
-                        vec![primary]
-                    }
-                };
-
-                let mut last_err: Option<String> = None;
-                let mut output_opt = None;
-                let mut chosen_cmd = String::new();
-                let mut chosen_args: Vec<String> = vec![];
-                for (cmd, args) in &attempts {
-                    match crate::util::process_manager::create_command(cmd).args(args).output() {
-                        Ok(out) => {
-                            if out.status.success() {
-                                chosen_cmd = cmd.clone();
-                                chosen_args = args.clone();
-                                output_opt = Some(out);
-                                break;
-                            } else {
-                                last_err = Some(format!(
-                                    "{} exit={:?} stderr={}",
-                                    cmd,
-                                    out.status.code(),
-                                    String::from_utf8_lossy(&out.stderr).trim()
-                                ));
-                            }
-                        }
-                        Err(e) => {
-                            last_err = Some(format!("spawn {}: {}", cmd, e));
-                        }
-                    }
-                }
-                let _ = chosen_args;
-                let output = output_opt.ok_or_else(|| {
-                    BitFunError::tool(format!(
-                        "open_app failed for '{}' across {} strategies: {} (host_error: {:?})",
-                        app_name,
-                        attempts.len(),
-                        last_err.as_deref().unwrap_or("(no error)"),
-                        host_error
-                    ))
+                let provider = LocalSystemProvider::new();
+                let outcome = provider.open_app_shell(app_name).map_err(|e| {
+                    BitFunError::tool(format!("{} (host_error: {:?})", e.message(), host_error))
                 })?;
-
-                if output.status.success() {
-                    let warning = host_error.map(|e| {
-                        format!("computer_use_host open_app failed; shell fallback succeeded: {}", e)
-                    });
-                    Ok(vec![ToolResult::ok(
-                        json!({
-                            "launched": true,
-                            "app": app_name,
-                            "method": method,
-                            "via_command": chosen_cmd,
-                            "host_attempted": host_attempted,
-                            "warning": warning,
-                        }),
-                        Some(format!("Opened {} via {}", app_name, chosen_cmd)),
-                    )])
-                } else {
-                    let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-                    Err(BitFunError::tool(format!(
-                        "open_app failed for '{}'. host_attempted={}, host_error={:?}, last_command='{}', stderr='{}'",
-                        app_name, host_attempted, host_error, chosen_cmd, stderr
-                    )))
-                }
+                let warning = host_error.map(|e| {
+                    format!("computer_use_host open_app failed; shell fallback succeeded: {}", e)
+                });
+                Ok(vec![ToolResult::ok(
+                    json!({
+                        "launched": true,
+                        "app": app_name,
+                        "method": method,
+                        "via_command": outcome.via_command,
+                        "host_attempted": host_attempted,
+                        "warning": warning,
+                    }),
+                    Some(format!("Opened {} via {}", app_name, outcome.via_command)),
+                )])
             }
             "run_script" => {
                 let script = params
@@ -1503,236 +1386,37 @@ impl ComputerUseActions {
                     .get("timeout_ms")
                     .and_then(|v| v.as_u64())
                     .filter(|value| *value > 0);
-                // Phase 4: keep output payloads bounded — model context is
-                // expensive and most scripts are happy with the head + tail.
-                let max_output_bytes = params
-                    .get("max_output_bytes")
-                    .and_then(|v| v.as_u64())
-                    .unwrap_or(16 * 1024)
-                    .clamp(1024, 256 * 1024) as usize;
-
-                let (program, args) = match script_type {
-                    "applescript" => {
-                        #[cfg(target_os = "macos")]
-                        {
-                            (
-                                "/usr/bin/osascript".to_string(),
-                                vec!["-e".to_string(), script.to_string()],
-                            )
-                        }
-                        #[cfg(not(target_os = "macos"))]
-                        {
-                            let _ = script;
-                            return Ok(err_response(
-                                "system",
-                                "run_script",
-                                ControlHubError::new(
-                                    ErrorCode::NotAvailable,
-                                    "AppleScript is only available on macOS",
-                                )
-                                .with_hint("Use script_type='shell' (sh on Unix, PowerShell on Windows) or script_type='powershell'/'bash'"),
-                            ));
-                        }
-                    }
-                    // The "shell" alias picks the OS's *default* shell so the
-                    // model can stay platform-agnostic. On Windows we now
-                    // route to PowerShell rather than cmd.exe to avoid the
-                    // GBK/CP936 stdout encoding nightmare and to give the
-                    // model a consistent surface area.
-                    "shell" => {
-                        #[cfg(target_os = "windows")]
-                        {
-                            powershell_invocation(script)
-                        }
-                        #[cfg(not(target_os = "windows"))]
-                        {
-                            (
-                                "sh".to_string(),
-                                vec!["-c".to_string(), script.to_string()],
-                            )
-                        }
-                    }
-                    "bash" => {
-                        // Bash is universally requested but not always on
-                        // PATH (Windows without WSL/git-bash). Detect and
-                        // surface a structured NotAvailable instead of a
-                        // confusing spawn-failure error.
-                        if !which_exists("bash") {
-                            return Ok(err_response(
-                                "system",
-                                "run_script",
-                                ControlHubError::new(
-                                    ErrorCode::NotAvailable,
-                                    "bash is not on PATH",
-                                )
-                                .with_hint("Install Git for Windows / WSL, or use script_type='shell' / 'powershell' / 'cmd'"),
-                            ));
-                        }
-                        (
-                            "bash".to_string(),
-                            vec!["-c".to_string(), script.to_string()],
-                        )
-                    }
-                    "powershell" => {
-                        // Prefer pwsh (PowerShell 7+, cross-platform) when
-                        // available; fall back to legacy Windows powershell.
-                        let prog = if which_exists("pwsh") {
-                            "pwsh"
-                        } else if which_exists("powershell") {
-                            "powershell"
-                        } else {
-                            return Ok(err_response(
-                                "system",
-                                "run_script",
-                                ControlHubError::new(
-                                    ErrorCode::NotAvailable,
-                                    "Neither pwsh nor powershell are on PATH",
-                                )
-                                .with_hint("Install PowerShell, or use script_type='shell' / 'bash'"),
-                            ));
-                        };
-                        (
-                            prog.to_string(),
-                            vec![
-                                "-NoProfile".to_string(),
-                                "-NonInteractive".to_string(),
-                                // -OutputEncoding utf8 is set inside the script
-                                // wrapper below for consistent stdout handling.
-                                "-Command".to_string(),
-                                format!(
-                                    "[Console]::OutputEncoding=[Text.Encoding]::UTF8; {}",
-                                    script
-                                ),
-                            ],
-                        )
-                    }
-                    "cmd" => {
-                        #[cfg(target_os = "windows")]
-                        {
-                            // Force code-page 65001 (UTF-8) before running the
-                            // user's script so stdout matches what we decode.
-                            (
-                                "cmd".to_string(),
-                                vec![
-                                    "/U".to_string(),
-                                    "/C".to_string(),
-                                    format!("chcp 65001>nul && {}", script),
-                                ],
-                            )
-                        }
-                        #[cfg(not(target_os = "windows"))]
-                        {
-                            return Ok(err_response(
-                                "system",
-                                "run_script",
-                                ControlHubError::new(
-                                    ErrorCode::NotAvailable,
-                                    "script_type='cmd' is only available on Windows",
-                                )
-                                .with_hint("Use script_type='shell' / 'bash' / 'powershell'"),
-                            ));
-                        }
-                    }
-                    other => {
-                        return Err(BitFunError::tool(format!(
-                            "Unknown script_type: '{}'. Valid: applescript (macOS), shell (OS default), bash, powershell, cmd (Windows)",
-                            other
-                        )))
-                    }
+                let max_output_bytes = params.get("max_output_bytes").and_then(|v| v.as_u64());
+                let provider = LocalSystemProvider::new();
+                let output = match provider
+                    .run_script(RunScriptRequest {
+                        script,
+                        script_type,
+                        timeout_ms,
+                        max_output_bytes,
+                    })
+                    .await
+                {
+                    Ok(output) => output,
+                    Err(error) => return map_run_script_error(error),
                 };
 
-                // Use tokio::process so that on timeout we can actually KILL
-                // the child process. The previous implementation wrapped
-                // `std::process::Command::output()` in `spawn_blocking` +
-                // `tokio::time::timeout`; on timeout the `timeout` future
-                // returned, but the spawn_blocking thread kept blocking on
-                // the still-running child, leaking a thread + process per
-                // hung script.
-                let started = std::time::Instant::now();
-                let child = process_manager::create_tokio_command(&program)
-                    .args(&args)
-                    .stdout(std::process::Stdio::piped())
-                    .stderr(std::process::Stdio::piped())
-                    .kill_on_drop(true)
-                    .spawn()
-                    .map_err(|e| {
-                        BitFunError::tool(format!(
-                            "Failed to spawn run_script ({}): {}",
-                            script_type, e
-                        ))
-                    })?;
-
-                let wait = child.wait_with_output();
-                let output = if let Some(timeout_ms) = timeout_ms {
-                    match tokio::time::timeout(std::time::Duration::from_millis(timeout_ms), wait)
-                        .await
-                    {
-                        Err(_) => {
-                            // Best-effort kill. `kill_on_drop(true)` above also
-                            // ensures the OS reaps the process when `child`
-                            // drops, but we issue an explicit SIGKILL first so
-                            // it terminates immediately rather than after the
-                            // tokio task tear-down race.
-                            // NOTE: `wait_with_output` consumed `child`, so we
-                            // can no longer call `child.kill()` directly here;
-                            // the `kill_on_drop` flag handles it for us.
-                            return Ok(err_response(
-                                "system",
-                                "run_script",
-                                ControlHubError::new(
-                                    ErrorCode::Timeout,
-                                    format!(
-                                        "run_script timed out after {} ms (script_type={}); child process killed",
-                                        timeout_ms, script_type
-                                    ),
-                                )
-                                .with_hint(
-                                    "Increase 'timeout_ms', set it to 0, or omit it to wait without a timeout",
-                                ),
-                            ));
-                        }
-                        Ok(Err(e)) => {
-                            return Err(BitFunError::tool(format!(
-                                "Failed to wait for run_script ({}): {}",
-                                script_type, e
-                            )));
-                        }
-                        Ok(Ok(o)) => o,
-                    }
-                } else {
-                    match wait.await {
-                        Ok(o) => o,
-                        Err(e) => {
-                            return Err(BitFunError::tool(format!(
-                                "Failed to wait for run_script ({}): {}",
-                                script_type, e
-                            )));
-                        }
-                    }
-                };
-
-                let elapsed_ms = elapsed_ms_u64(started);
-                let stdout_full = String::from_utf8_lossy(&output.stdout).to_string();
-                let stderr_full = String::from_utf8_lossy(&output.stderr).to_string();
-                let (stdout, stdout_truncated) = truncate_with_marker(&stdout_full, max_output_bytes);
-                let (stderr, stderr_truncated) = truncate_with_marker(&stderr_full, max_output_bytes);
-
-                if output.status.success() {
+                if output.success {
                     Ok(vec![ToolResult::ok(
                         json!({
                             "success": true,
-                            "output": stdout,
-                            "stderr": stderr,
-                            "stdout_truncated": stdout_truncated,
-                            "stderr_truncated": stderr_truncated,
-                            "exit_code": output.status.code(),
-                            "elapsed_ms": elapsed_ms,
+                            "output": output.stdout,
+                            "stderr": output.stderr,
+                            "stdout_truncated": output.stdout_truncated,
+                            "stderr_truncated": output.stderr_truncated,
+                            "exit_code": output.exit_code,
+                            "elapsed_ms": output.elapsed_ms,
                             "script_type": script_type,
                         }),
-                        Some(if stdout.is_empty() {
-                            format!("Script executed in {} ms", elapsed_ms)
+                        Some(if output.stdout.is_empty() {
+                            format!("Script executed in {} ms", output.elapsed_ms)
                         } else {
-                            stdout.lines().take(1).collect::<String>()
+                            output.stdout.lines().take(1).collect::<String>()
                         }),
                     )])
                 } else {
@@ -1743,71 +1427,44 @@ impl ComputerUseActions {
                             ErrorCode::Internal,
                             format!(
                                 "Script exited with {:?}: {}",
-                                output.status.code(),
-                                stderr.lines().next().unwrap_or("(no stderr)")
+                                output.exit_code,
+                                output.stderr.lines().next().unwrap_or("(no stderr)")
                             ),
                         )
                         .with_hints([
-                            format!("stderr={}", stderr),
-                            format!("elapsed_ms={}", elapsed_ms),
+                            format!("stderr={}", output.stderr),
+                            format!("elapsed_ms={}", output.elapsed_ms),
                         ]),
                     ))
                 }
             }
             "get_os_info" => {
-                let os = std::env::consts::OS;
-                let arch = std::env::consts::ARCH;
-                // Phase 4: include OS version + hostname when available so
-                // the model can adapt platform-specific paths / commands.
+                let local = LocalSystemProvider::new().system_info();
                 let mut info = json!({
-                    "os": os,
-                    "arch": arch,
-                    "rust_target_family": std::env::consts::FAMILY,
+                    "os": local.os,
+                    "arch": local.arch,
+                    "rust_target_family": local.rust_target_family,
                 });
-                if let Some(v) = read_os_version() {
+                if let Some(v) = local.os_version {
                     info["os_version"] = json!(v);
                 }
-                if let Ok(host) = hostname() {
+                if let Some(host) = local.hostname {
                     info["hostname"] = json!(host);
                 }
-                // Linux-only: surface display server (X11 / Wayland) and the
-                // current desktop environment so the model can pick the right
-                // clipboard helper / window manipulation strategy without a
-                // separate `run_script` round-trip.
-                #[cfg(target_os = "linux")]
-                {
-                    let (display_server, desktop_env) = linux_session_info();
-                    if let Some(s) = display_server {
-                        info["display_server"] = json!(s);
-                    }
-                    if let Some(d) = desktop_env {
-                        info["desktop_environment"] = json!(d);
-                    }
+                if let Some(s) = local.display_server {
+                    info["display_server"] = json!(s);
                 }
-                // The set of `script_type` values the host can actually run.
-                // Discoverability win: model no longer has to spawn a doomed
-                // run_script call to learn that bash is missing on Windows.
-                let mut script_types = vec!["shell"];
-                if cfg!(target_os = "macos") {
-                    script_types.push("applescript");
+                if let Some(d) = local.desktop_environment {
+                    info["desktop_environment"] = json!(d);
                 }
-                if which_exists("bash") {
-                    script_types.push("bash");
-                }
-                if which_exists("pwsh") || which_exists("powershell") {
-                    script_types.push("powershell");
-                }
-                if cfg!(target_os = "windows") {
-                    script_types.push("cmd");
-                }
-                info["script_types"] = json!(script_types);
+                info["script_types"] = json!(local.script_types);
                 Ok(vec![ToolResult::ok(
                     info.clone(),
                     Some(format!(
                         "{} {} ({})",
-                        os,
+                        local.os,
                         info.get("os_version").and_then(|v| v.as_str()).unwrap_or(""),
-                        arch
+                        local.arch
                     )),
                 )])
             }
@@ -1824,7 +1481,7 @@ impl ComputerUseActions {
                     .unwrap_or(64 * 1024)
                     .clamp(64, 1024 * 1024);
 
-                match clipboard_read().await {
+                match LocalSystemProvider::new().clipboard_read_text().await {
                     Ok(text) => {
                         let (truncated, was_truncated) = truncate_with_marker(&text, max_bytes);
                         let len = text.len();
@@ -1837,15 +1494,7 @@ impl ComputerUseActions {
                             Some(format!("{} bytes on clipboard", len)),
                         )])
                     }
-                    Err(e) => Ok(err_response(
-                        "system",
-                        "clipboard_get",
-                        ControlHubError::new(
-                            ErrorCode::NotAvailable,
-                            format!("Clipboard read failed: {}", e),
-                        )
-                        .with_hints(linux_clipboard_install_hints()),
-                    )),
+                    Err(e) => Ok(local_system_error_response("system", "clipboard_get", e)),
                 }
             }
 
@@ -1856,7 +1505,7 @@ impl ComputerUseActions {
                 let text = params.get("text").and_then(|v| v.as_str()).ok_or_else(|| {
                     BitFunError::tool("clipboard_set requires 'text'".to_string())
                 })?;
-                match clipboard_write(text).await {
+                match LocalSystemProvider::new().clipboard_write_text(text).await {
                     Ok(()) => Ok(vec![ToolResult::ok(
                         json!({
                             "success": true,
@@ -1864,15 +1513,7 @@ impl ComputerUseActions {
                         }),
                         Some(format!("Wrote {} bytes to clipboard", text.len())),
                     )]),
-                    Err(e) => Ok(err_response(
-                        "system",
-                        "clipboard_set",
-                        ControlHubError::new(
-                            ErrorCode::NotAvailable,
-                            format!("Clipboard write failed: {}", e),
-                        )
-                        .with_hints(linux_clipboard_install_hints()),
-                    )),
+                    Err(e) => Ok(local_system_error_response("system", "clipboard_set", e)),
                 }
             }
 
@@ -1886,65 +1527,12 @@ impl ComputerUseActions {
                     .get("url")
                     .and_then(|v| v.as_str())
                     .ok_or_else(|| BitFunError::tool("open_url requires 'url'".to_string()))?;
-                if !(url.starts_with("http://")
-                    || url.starts_with("https://")
-                    || url.starts_with("file://")
-                    || url.starts_with("mailto:"))
-                {
-                    return Ok(err_response(
-                        "system",
-                        "open_url",
-                        ControlHubError::new(
-                            ErrorCode::InvalidParams,
-                            format!("Refusing to open URL with unsupported scheme: {}", url),
-                        )
-                        .with_hint(
-                            "Pass an http(s)://, file://, or mailto: URL. Use 'open_file' for local paths without a scheme.",
-                        ),
-                    ));
-                }
-                // NOTE: do NOT reuse platform_open_command — that helper
-                // is for *apps* (uses `open -a` on macOS) and would treat
-                // the URL as an application name, failing immediately.
-                //
-                // Windows: must NOT route through `cmd /C start "" <url>`.
-                // `cmd` interprets `&`, `^`, `%`, `|` in the URL — so a query
-                // string like `?a=1&b=2` gets the second arg dropped, and
-                // long URLs may be silently truncated. Use rundll32 with the
-                // URL protocol handler so the URL is passed verbatim and
-                // routed through the same default-handler resolution Windows
-                // uses for "Open in Browser" shell verbs.
-                let (program, args) = match std::env::consts::OS {
-                    "macos" => ("open".to_string(), vec![url.to_string()]),
-                    "windows" => (
-                        "rundll32".to_string(),
-                        vec![
-                            "url.dll,FileProtocolHandler".to_string(),
-                            url.to_string(),
-                        ],
-                    ),
-                    _ => ("xdg-open".to_string(), vec![url.to_string()]),
-                };
-                let status = process_manager::create_command(&program)
-                    .args(&args)
-                    .status()
-                    .map_err(|e| {
-                        BitFunError::tool(format!("Failed to spawn '{}': {}", program, e))
-                    })?;
-                if status.success() {
-                    Ok(vec![ToolResult::ok(
-                        json!({ "opened": true, "url": url, "method": program }),
+                match LocalSystemProvider::new().open_url(url) {
+                    Ok(outcome) => Ok(vec![ToolResult::ok(
+                        json!({ "opened": true, "url": url, "method": outcome.method }),
                         Some(format!("Opened {} in default handler", url)),
-                    )])
-                } else {
-                    Ok(err_response(
-                        "system",
-                        "open_url",
-                        ControlHubError::new(
-                            ErrorCode::Internal,
-                            format!("'{}' exited with {:?}", program, status.code()),
-                        ),
-                    ))
+                    )]),
+                    Err(e) => Ok(local_system_error_response("system", "open_url", e)),
                 }
             }
 
@@ -1957,65 +1545,20 @@ impl ComputerUseActions {
                 })?;
                 let app_name = params.get("app").and_then(|v| v.as_str());
 
-                let path = std::path::Path::new(path_str);
-                if !path.exists() {
-                    return Ok(err_response(
-                        "system",
-                        "open_file",
-                        ControlHubError::new(
-                            ErrorCode::NotFound,
-                            format!("File does not exist: {}", path_str),
-                        )
-                        .with_hint("Check the absolute path; ~ is not expanded"),
-                    ));
-                }
-
-                let (program, args) = match (std::env::consts::OS, app_name) {
-                    ("macos", Some(app)) => (
-                        "open".to_string(),
-                        vec!["-a".to_string(), app.to_string(), path_str.to_string()],
-                    ),
-                    ("macos", None) => ("open".to_string(), vec![path_str.to_string()]),
-                    // Windows file open: same rundll32 dance as open_url so
-                    // paths with `&` / `%` survive intact when cmd would have
-                    // mangled them. ShellExec_RunDLL also accepts file paths.
-                    ("windows", _) => (
-                        "rundll32".to_string(),
-                        vec![
-                            "url.dll,FileProtocolHandler".to_string(),
-                            path_str.to_string(),
-                        ],
-                    ),
-                    _ => ("xdg-open".to_string(), vec![path_str.to_string()]),
-                };
-                let status = process_manager::create_command(&program)
-                    .args(&args)
-                    .status()
-                    .map_err(|e| {
-                        BitFunError::tool(format!("Failed to spawn '{}': {}", program, e))
-                    })?;
-                if status.success() {
-                    Ok(vec![ToolResult::ok(
+                match LocalSystemProvider::new().open_file(path_str, app_name) {
+                    Ok(outcome) => Ok(vec![ToolResult::ok(
                         json!({
                             "opened": true,
                             "path": path_str,
                             "with_app": app_name,
-                            "method": program,
+                            "method": outcome.method,
                         }),
                         Some(match app_name {
                             Some(a) => format!("Opened {} with {}", path_str, a),
                             None => format!("Opened {} with default handler", path_str),
                         }),
-                    )])
-                } else {
-                    Ok(err_response(
-                        "system",
-                        "open_file",
-                        ControlHubError::new(
-                            ErrorCode::Internal,
-                            format!("'{}' exited with {:?}", program, status.code()),
-                        ),
-                    ))
+                    )]),
+                    Err(e) => Ok(local_system_error_response("system", "open_file", e)),
                 }
             }
 
@@ -2026,361 +1569,44 @@ impl ComputerUseActions {
         }
     }
 }
-/// Truncate `s` to at most `max_bytes`, appending an explicit marker so the
-/// model can see that data was dropped (and how much). Returns
-/// `(truncated_string, was_truncated)`.
-pub(crate) fn truncate_with_marker(s: &str, max_bytes: usize) -> (String, bool) {
-    if s.len() <= max_bytes {
-        return (s.to_string(), false);
+fn local_system_error_response(
+    domain: &'static str,
+    action: &'static str,
+    error: LocalSystemActionError,
+) -> Vec<ToolResult> {
+    let mut control_error =
+        ControlHubError::new(error_code_from_local(error.kind()), error.message());
+    if !error.hints().is_empty() {
+        control_error = control_error.with_hints(error.hints().to_vec());
     }
-    let head_n = max_bytes.saturating_sub(64);
-    let head = safe_str_slice(s, head_n);
-    let omitted = s.len().saturating_sub(head_n);
-    (
-        format!("{}\n... [{} bytes omitted] ...\n", head, omitted),
-        true,
-    )
-}
-/// Slice `s` to ≤ `n` bytes without splitting a UTF-8 codepoint.
-fn safe_str_slice(s: &str, n: usize) -> &str {
-    if n >= s.len() {
-        return s;
-    }
-    let mut cut = n;
-    while cut > 0 && !s.is_char_boundary(cut) {
-        cut -= 1;
-    }
-    &s[..cut]
+    err_response(domain, action, control_error)
 }
 
-/// Read a short OS version string. Best-effort: returns `None` on platforms
-/// where we can't determine it cheaply.
-fn read_os_version() -> Option<String> {
-    #[cfg(target_os = "macos")]
-    {
-        let out = std::process::Command::new("sw_vers")
-            .arg("-productVersion")
-            .output()
-            .ok()?;
-        let s = String::from_utf8_lossy(&out.stdout).trim().to_string();
-        if s.is_empty() {
-            None
-        } else {
-            Some(format!("macOS {}", s))
+fn map_run_script_error(error: LocalSystemActionError) -> BitFunResult<Vec<ToolResult>> {
+    match error.kind() {
+        LocalSystemActionErrorKind::NotAvailable | LocalSystemActionErrorKind::Timeout => {
+            Ok(local_system_error_response("system", "run_script", error))
         }
-    }
-    #[cfg(target_os = "windows")]
-    {
-        let out = crate::util::process_manager::create_command("cmd")
-            .args(["/C", "ver"])
-            .output()
-            .ok()?;
-        let s = String::from_utf8_lossy(&out.stdout).trim().to_string();
-        if s.is_empty() {
-            None
-        } else {
-            Some(s)
+        LocalSystemActionErrorKind::UnknownScriptType => {
+            Err(BitFunError::tool(error.message().to_string()))
         }
-    }
-    #[cfg(target_os = "linux")]
-    {
-        // /etc/os-release is the canonical lookup.
-        let txt = std::fs::read_to_string("/etc/os-release").ok()?;
-        for line in txt.lines() {
-            if let Some(rest) = line.strip_prefix("PRETTY_NAME=") {
-                return Some(rest.trim_matches('"').to_string());
-            }
+        LocalSystemActionErrorKind::InvalidParams
+        | LocalSystemActionErrorKind::NotFound
+        | LocalSystemActionErrorKind::Internal => {
+            Err(BitFunError::tool(error.message().to_string()))
         }
-        None
-    }
-    #[cfg(not(any(target_os = "macos", target_os = "windows", target_os = "linux")))]
-    {
-        None
     }
 }
 
-fn hostname() -> std::io::Result<String> {
-    // Prefer environment variables on each OS so we never have to spawn a
-    // subprocess for a value that's already in our address space, and so we
-    // never ingest a non-UTF-8 byte stream from `hostname.exe` on Windows
-    // running a CJK code page.
-    #[cfg(target_os = "windows")]
-    {
-        if let Ok(name) = std::env::var("COMPUTERNAME") {
-            if !name.is_empty() {
-                return Ok(name);
-            }
+fn error_code_from_local(kind: LocalSystemActionErrorKind) -> ErrorCode {
+    match kind {
+        LocalSystemActionErrorKind::InvalidParams => ErrorCode::InvalidParams,
+        LocalSystemActionErrorKind::NotAvailable => ErrorCode::NotAvailable,
+        LocalSystemActionErrorKind::NotFound => ErrorCode::NotFound,
+        LocalSystemActionErrorKind::Timeout => ErrorCode::Timeout,
+        LocalSystemActionErrorKind::Internal | LocalSystemActionErrorKind::UnknownScriptType => {
+            ErrorCode::Internal
         }
-    }
-    #[cfg(any(target_os = "linux", target_os = "macos"))]
-    {
-        if let Ok(name) = std::env::var("HOSTNAME") {
-            if !name.is_empty() {
-                return Ok(name);
-            }
-        }
-        if let Ok(bytes) = std::fs::read("/etc/hostname") {
-            let s = String::from_utf8_lossy(&bytes).trim().to_string();
-            if !s.is_empty() {
-                return Ok(s);
-            }
-        }
-    }
-    let out = process_manager::create_command("hostname").output()?;
-    Ok(String::from_utf8_lossy(&out.stdout).trim().to_string())
-}
-
-/// Cheap PATH lookup for an executable name. Used to decide between e.g.
-/// `pwsh` and `powershell`, or to surface a structured `NOT_AVAILABLE`
-/// error when the requested interpreter isn't installed.
-pub(crate) fn which_exists(name: &str) -> bool {
-    let paths = match std::env::var_os("PATH") {
-        Some(p) => p,
-        None => return false,
-    };
-    let exts: Vec<String> = if cfg!(target_os = "windows") {
-        std::env::var("PATHEXT")
-            .unwrap_or_else(|_| ".EXE;.BAT;.CMD;.COM".to_string())
-            .split(';')
-            .map(|s| s.to_string())
-            .collect()
-    } else {
-        vec![String::new()]
-    };
-    for dir in std::env::split_paths(&paths) {
-        for ext in &exts {
-            let mut candidate = dir.join(name);
-            if !ext.is_empty() {
-                let stem = candidate.file_name().map(|n| n.to_os_string());
-                if let Some(mut stem) = stem {
-                    stem.push(ext);
-                    candidate.set_file_name(stem);
-                }
-            }
-            if candidate.exists() {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-/// Build a `(program, args)` pair for invoking a PowerShell snippet on Windows
-/// with UTF-8 output forced. Centralised so the "shell" alias and an explicit
-/// `script_type='powershell'` produce the same encoding.
-#[cfg(target_os = "windows")]
-fn powershell_invocation(script: &str) -> (String, Vec<String>) {
-    let prog = if which_exists("pwsh") {
-        "pwsh"
-    } else {
-        "powershell"
-    };
-    (
-        prog.to_string(),
-        vec![
-            "-NoProfile".to_string(),
-            "-NonInteractive".to_string(),
-            "-Command".to_string(),
-            format!(
-                "[Console]::OutputEncoding=[Text.Encoding]::UTF8; {}",
-                script
-            ),
-        ],
-    )
-}
-
-/// Build OS-specific install hints for the clipboard helper. On Linux we
-/// inspect the session type so the suggestion matches what the user actually
-/// needs (Wayland users wasting time installing xclip is a real failure mode).
-pub(crate) fn linux_clipboard_install_hints() -> Vec<String> {
-    match std::env::consts::OS {
-        "linux" => {
-            #[cfg(target_os = "linux")]
-            {
-                let (server, _) = linux_session_info();
-                match server.as_deref() {
-                    Some("wayland") => vec![
-                        "Wayland session detected — install wl-clipboard (e.g. `sudo apt install wl-clipboard` / `sudo dnf install wl-clipboard`)".to_string(),
-                        "Fallback for XWayland apps: also install xclip or xsel".to_string(),
-                    ],
-                    Some("x11") | Some("tty") => vec![
-                        "X11 session detected — install xclip (`sudo apt install xclip`) or xsel (`sudo apt install xsel`)".to_string(),
-                    ],
-                    _ => vec![
-                        "Install wl-clipboard (Wayland) OR xclip/xsel (X11). Run `echo $XDG_SESSION_TYPE` to know which one applies.".to_string(),
-                    ],
-                }
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                vec!["Install wl-clipboard (Wayland) or xclip/xsel (X11)".to_string()]
-            }
-        }
-        _ => vec!["Make sure the system clipboard helper is available on this host".to_string()],
-    }
-}
-/// Best-effort detection of the Linux desktop session metadata (display
-/// server + desktop environment). Returns `(display_server, desktop_env)`,
-/// either of which may be `None` if the environment doesn't expose it.
-#[cfg(target_os = "linux")]
-pub(crate) fn linux_session_info() -> (Option<String>, Option<String>) {
-    let server = std::env::var("XDG_SESSION_TYPE")
-        .ok()
-        .filter(|s| !s.is_empty());
-    let de = std::env::var("XDG_CURRENT_DESKTOP")
-        .ok()
-        .or_else(|| std::env::var("DESKTOP_SESSION").ok())
-        .filter(|s| !s.is_empty());
-    (server, de)
-}
-
-/// Cross-platform clipboard read. Shells out to the canonical helper for
-/// the current OS so we don't pull in a heavyweight dependency for what is
-/// fundamentally a 1-line operation. Linux auto-detects Wayland → X11.
-async fn clipboard_read() -> Result<String, String> {
-    #[cfg(target_os = "macos")]
-    {
-        let out = process_manager::create_tokio_command("pbpaste")
-            .output()
-            .await
-            .map_err(|e| format!("spawn pbpaste: {}", e))?;
-        if !out.status.success() {
-            return Err(format!("pbpaste exit={:?}", out.status.code()));
-        }
-        Ok(String::from_utf8_lossy(&out.stdout).to_string())
-    }
-    #[cfg(target_os = "windows")]
-    {
-        let (program, args) = powershell_invocation("Get-Clipboard -Raw");
-        let out = process_manager::create_tokio_command(&program)
-            .args(&args)
-            .output()
-            .await
-            .map_err(|e| format!("spawn {}: {}", program, e))?;
-        if !out.status.success() {
-            return Err(format!("Get-Clipboard exit={:?}", out.status.code()));
-        }
-        // PowerShell appends CRLF; trim a single trailing newline so the
-        // returned text matches what the user actually copied.
-        let mut s = String::from_utf8_lossy(&out.stdout).to_string();
-        if s.ends_with("\r\n") {
-            s.truncate(s.len() - 2);
-        } else if s.ends_with('\n') {
-            s.truncate(s.len() - 1);
-        }
-        Ok(s)
-    }
-    #[cfg(target_os = "linux")]
-    {
-        // Wayland first (modern session), then X11 fallbacks.
-        let candidates: &[(&str, &[&str])] = if std::env::var("WAYLAND_DISPLAY").is_ok() {
-            &[
-                ("wl-paste", &["--no-newline"]),
-                ("xclip", &["-selection", "clipboard", "-o"]),
-                ("xsel", &["--clipboard", "--output"]),
-            ]
-        } else {
-            &[
-                ("xclip", &["-selection", "clipboard", "-o"]),
-                ("xsel", &["--clipboard", "--output"]),
-                ("wl-paste", &["--no-newline"]),
-            ]
-        };
-        for (bin, args) in candidates {
-            if let Ok(out) = process_manager::create_tokio_command(bin)
-                .args(*args)
-                .output()
-                .await
-            {
-                if out.status.success() {
-                    return Ok(String::from_utf8_lossy(&out.stdout).to_string());
-                }
-            }
-        }
-        Err("no clipboard helper found (install wl-clipboard, xclip, or xsel)".to_string())
-    }
-    #[cfg(not(any(target_os = "macos", target_os = "windows", target_os = "linux")))]
-    {
-        Err("clipboard not implemented for this OS".to_string())
-    }
-}
-
-/// Cross-platform clipboard write. Streams `text` into the helper's stdin
-/// rather than embedding it in argv so newlines / quotes / shell metachars
-/// are preserved verbatim.
-async fn clipboard_write(text: &str) -> Result<(), String> {
-    use tokio::io::AsyncWriteExt;
-
-    async fn pipe(bin: &str, args: &[&str], text: &str) -> Result<(), String> {
-        let mut child = process_manager::create_tokio_command(bin)
-            .args(args)
-            .stdin(std::process::Stdio::piped())
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::piped())
-            .spawn()
-            .map_err(|e| format!("spawn {}: {}", bin, e))?;
-        if let Some(mut stdin) = child.stdin.take() {
-            stdin
-                .write_all(text.as_bytes())
-                .await
-                .map_err(|e| format!("write {} stdin: {}", bin, e))?;
-        }
-        let out = child
-            .wait_with_output()
-            .await
-            .map_err(|e| format!("wait {}: {}", bin, e))?;
-        if !out.status.success() {
-            return Err(format!("{} exit={:?}", bin, out.status.code()));
-        }
-        Ok(())
-    }
-
-    #[cfg(target_os = "macos")]
-    {
-        pipe("pbcopy", &[], text).await
-    }
-    #[cfg(target_os = "windows")]
-    {
-        // PowerShell's Set-Clipboard reads from the pipeline; pipe text in
-        // via stdin to preserve binary fidelity.
-        pipe(
-            "powershell",
-            &["-NoProfile", "-Command", "$input | Set-Clipboard"],
-            text,
-        )
-        .await
-    }
-    #[cfg(target_os = "linux")]
-    {
-        let candidates: &[(&str, &[&str])] = if std::env::var("WAYLAND_DISPLAY").is_ok() {
-            &[
-                ("wl-copy", &[]),
-                ("xclip", &["-selection", "clipboard"]),
-                ("xsel", &["--clipboard", "--input"]),
-            ]
-        } else {
-            &[
-                ("xclip", &["-selection", "clipboard"]),
-                ("xsel", &["--clipboard", "--input"]),
-                ("wl-copy", &[]),
-            ]
-        };
-        let mut last_err = String::new();
-        for (bin, args) in candidates {
-            match pipe(bin, args, text).await {
-                Ok(()) => return Ok(()),
-                Err(e) => last_err = e,
-            }
-        }
-        Err(format!(
-            "no clipboard helper succeeded (install wl-clipboard, xclip, or xsel): {}",
-            last_err
-        ))
-    }
-    #[cfg(not(any(target_os = "macos", target_os = "windows", target_os = "linux")))]
-    {
-        let _ = text;
-        Err("clipboard not implemented for this OS".to_string())
     }
 }
